Add files via upload
In this project I have used Pytorch to create a reinforcement learning model that learns to play the classic game of Snake where the objective is to create the highest score possible by eating "fruits". I have created a small NN for this which consists of an Input layer of size 11, a hidden layer of size 256 and an output layer of size 3. The size of the hidden layer has been chosen at random and can be changed but the size of the input layer represents 11 states based on which the snake is supposed to move and the size of the input layer represents the 3 directions the snake can move in at any given point.